# backend/src/AosAdjutant.Api/Features/Factions/FactionUnitController.cs
using AosAdjutant.Api.Common;
using AosAdjutant.Api.Features.Units;
using Microsoft.AspNetCore.Mvc;

namespace AosAdjutant.Api.Features.Factions;

[Route("api/factions/{factionId:int}/units")]
[ApiController]
[Tags("Units")]
public sealed class FactionUnitController(UnitService unitService) : ControllerBase
{
    [HttpPost]
    [EndpointSummary("Create a unit under a faction")]
    [ProducesResponseType<UnitResponseDto>(StatusCodes.Status201Created)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status404NotFound)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status409Conflict)]
    public async Task<ActionResult<UnitResponseDto>> CreateUnit(
        [FromRoute] int factionId,
        [FromBody] CreateUnitDto unitData
    )
    {
        var unitResult = await unitService.CreateUnit(factionId, unitData);
        return unitResult.Match(
            u =>
                CreatedAtAction(
                    nameof(UnitController.GetUnit),
                    "Unit",
                    new { unitId = u.UnitId },
                    new UnitResponseDto(
                        u.UnitId,
                        u.Name,
                        u.Health,
                        u.Move,
                        u.Save,
                        u.Control,
                        u.WardSave,
                        u.FactionId,
                        u.Version
                    )
                ),
            this.ApiProblem
        );
    }

    [HttpGet]
    [EndpointSummary("Get all units for a faction")]
    [ProducesResponseType<PaginatedResponse<UnitResponseDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<PaginatedResponse<UnitResponseDto>>> GetUnits(
        [FromRoute] int factionId,
        [FromQuery] UnitQuery unitQuery
    )
    {
        var unitsResult = await unitService.GetFactionUnits(factionId, unitQuery);
        return unitsResult.Match(
            units =>
                Ok(
                    units.Map(u => new UnitResponseDto(
                        u.UnitId,
                        u.Name,
                        u.Health,
                        u.Move,
                        u.Save,
                        u.Control,
                        u.WardSave,
                        u.FactionId,
                        u.Version
                    ))
                ),
            this.ApiProblem
        );
    }
}