# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AosAdjutant.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/units/{unitId}/abilities": {
      "post": {
        "tags": [
          "Units"
        ],
        "summary": "Create an ability for a unit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Units"
        ],
        "summary": "Get all abilities for a unit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Phase",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Phase"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AbilitySortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/factions/{factionId}/battle-formations": {
      "post": {
        "tags": [
          "Battle Formations"
        ],
        "summary": "Create a battle formation under a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBattleFormationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBattleFormationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBattleFormationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BattleFormationResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BattleFormationResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BattleFormationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Battle Formations"
        ],
        "summary": "Get all battle formations for a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/BattleFormationSortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfBattleFormationResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfBattleFormationResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfBattleFormationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/factions/{factionId}/units": {
      "post": {
        "tags": [
          "Units"
        ],
        "summary": "Create a unit under a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Units"
        ],
        "summary": "Get all units for a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/UnitSortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfUnitResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfUnitResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfUnitResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbilityResponseDto": {
        "required": [
          "abilityId",
          "name",
          "reaction",
          "declaration",
          "effect",
          "phase",
          "restriction",
          "turn",
          "version"
        ],
        "type": "object",
        "properties": {
          "abilityId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "reaction": {
            "type": [
              "null",
              "string"
            ]
          },
          "declaration": {
            "type": [
              "null",
              "string"
            ]
          },
          "effect": {
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "restriction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Restriction"
              }
            ]
          },
          "turn": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Turn"
              }
            ]
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          }
        }
      },
      "AbilitySortBy": {
        "enum": [
          "Name",
          "Phase",
          null
        ]
      },
...
      "BattleFormationResponseDto": {
        "required": [
          "battleFormationId",
          "name",
          "factionId",
          "version"
        ],
        "type": "object",
        "properties": {
          "battleFormationId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "factionId": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          }
        }
      },
      "BattleFormationSortBy": {
        "enum": [
          "Name",
          null
        ]
      },
...
      "CreateAbilityDto": {
        "required": [
          "name",
          "effect",
          "phase"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "reaction": {
            "maxLength": 100,
            "minLength": 1,
            "type": [
              "null",
              "string"
            ]
          },
          "declaration": {
            "maxLength": 100,
            "minLength": 1,
            "type": [
              "null",
              "string"
            ]
          },
          "effect": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "restriction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Restriction"
              }
            ]
          },
          "turn": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Turn"
              }
            ]
          }
        }
      },
...
      "CreateBattleFormationDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          }
        }
      },
...
      "CreateUnitDto": {
        "required": [
          "name",
          "health",
          "move",
          "save",
          "control"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "health": {
            "type": "integer",
            "format": "int32"
          },
          "move": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "save": {
            "type": "integer",
            "format": "int32"
          },
          "control": {
            "type": "integer",
            "format": "int32"
          },
          "wardSave": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "PaginatedResponseOfAbilityResponseDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbilityResponseDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PaginatedResponseOfBattleFormationResponseDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BattleFormationResponseDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PaginatedResponseOfUnitResponseDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitResponseDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Phase": {
        "enum": [
          "Deployment",
          "Start",
          "Hero",
          "Movement",
          "Shooting",
          "Charge",
          "Combat",
          "End",
          "Passive"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "Restriction": {
        "enum": [
          "OnceTurnArmy",
          "OnceRoundArmy",
          "OnceBattleArmy",
          "OnceRound",
          "OnceBattle",
          null
        ]
      },
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ]
      },
      "Turn": {
        "enum": [
          "YourTurn",
          "EnemyTurn",
          "AnyTurn",
          null
        ]
      },
      "UnitResponseDto": {
        "required": [
          "unitId",
          "name",
          "health",
          "move",
          "save",
          "control",
          "wardSave",
          "factionId",
          "version"
        ],
        "type": "object",
        "properties": {
          "unitId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "health": {
            "type": "integer",
            "format": "int32"
          },
          "move": {
            "type": "string"
          },
          "save": {
            "type": "integer",
            "format": "int32"
          },
          "control": {
            "type": "integer",
            "format": "int32"
          },
          "wardSave": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "factionId": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          }
        }
      },
      "UnitSortBy": {
        "enum": [
          "Name",
          null
        ]
      },
...
    }
  },
...
}